+++ src/sql/workbench/parts/notebook/cellViews/textCell.component.ts
@@ -25,16 +25,10 @@ import { ISanitizer, defaultSanitizer } from 'sql/workbench/parts/notebook/outpu
 import { NotebookModel } from 'sql/workbench/parts/notebook/models/notebookModel';
 import { CellToggleMoreActions } from 'sql/workbench/parts/notebook/cellToggleMoreActions';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { toDisposable, IDisposable, dispose } from 'vs/base/common/lifecycle';
-import { RenderOptions } from 'vs/base/browser/htmlContentRenderer';
-import { IMarkdownString, removeMarkdownEscapes } from 'vs/base/common/htmlContent';
+import { toDisposable } from 'vs/base/common/lifecycle';
 import { IMarkdownRenderResult } from 'vs/editor/contrib/markdown/markdownRenderer';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
-import { onUnexpectedError } from 'vs/base/common/errors';
-import marked = require('vs/base/common/marked/marked');
-import { defaultGenerator } from 'vs/base/common/idGenerator';
-import { revive } from 'vs/base/common/marshalling';
-import * as fs from 'fs';
+import { NotebookMarkdownRenderer } from 'sql/workbench/parts/notebook/outputs/notebookMarkdown';
 
 export const TEXT_SELECTOR: string = 'text-cell-component';
 const USER_SELECT_CLASS = 'actionselect';
@@ -84,7 +78,7 @@ export class TextCellComponent extends CellView implements OnInit, OnChanges {
 	public readonly onDidClickLink = this._onDidClickLink.event;
 	private _cellToggleMoreActions: CellToggleMoreActions;
 	private _hover: boolean;
-	private markdownRenderer: MarkdownRenderer;
+	private markdownRenderer: NotebookMarkdownRenderer;
 	private markdownResult: IMarkdownRenderResult;
 
 	constructor(
@@ -100,7 +94,7 @@ export class TextCellComponent extends CellView implements OnInit, OnChanges {
 		super();
 		this.isEditMode = true;
 		this._cellToggleMoreActions = this._instantiationService.createInstance(CellToggleMoreActions);
-		this.markdownRenderer = this._instantiationService.createInstance(MarkdownRenderer);
+		this.markdownRenderer = this._instantiationService.createInstance(NotebookMarkdownRenderer);
 		this._register(toDisposable(() => {
 			if (this.markdownResult) {
 				this.markdownResult.dispose();
@@ -176,8 +170,7 @@ export class TextCellComponent extends CellView implements OnInit, OnChanges {
 				this._content = this.cellModel.source;
 			}
 
-			if (this.useSimpleMarkdown) {
-				let time0 = performance.now();
+			if (this.useInProcMarkdown) {
 				let uri = Object.create(null);
 				this.markdownRenderer.setNotebookURI(this.cellModel.notebookModel.notebookUri);
 				this.markdownResult = this.markdownRenderer.render({
@@ -188,25 +181,20 @@ export class TextCellComponent extends CellView implements OnInit, OnChanges {
 				let outputElement = <HTMLElement>this.output.nativeElement;
 				DOM.clearNode(outputElement);
 				DOM.append(outputElement, this.markdownResult.element);
-				let elapsed = performance.now() - time0;
-				console.log(`time to render simply is ${elapsed}ms`);
 			} else {
-				let time0 = performance.now();
 				this._commandService.executeCommand<string>('notebook.showPreview', this.cellModel.notebookModel.notebookUri, this._content).then((htmlcontent) => {
 					htmlcontent = this.convertVscodeResourceToFileInSubDirectories(htmlcontent);
 					htmlcontent = this.sanitizeContent(htmlcontent);
 					let outputElement = <HTMLElement>this.output.nativeElement;
 					outputElement.innerHTML = htmlcontent;
 					this.setLoading(false);
-					let elapsed = performance.now() - time0;
-					console.log(`time to render over extension host is ${elapsed}ms`);
 				});
 			}
 		}
 	}
 
-	private get useSimpleMarkdown(): boolean {
-		return this.configurationService.getValue('notebook.useSimpleMarkdown');
+	private get useInProcMarkdown(): boolean {
+		return this.configurationService.getValue('notebook.useInProcMarkdown');
 	}
 
 
@@ -294,271 +282,4 @@ export class TextCellComponent extends CellView implements OnInit, OnChanges {
 	protected toggleMoreActionsButton(isActiveOrHovered: boolean) {
 		this._cellToggleMoreActions.toggleVisible(!isActiveOrHovered);
 	}
-}
-
-class MarkdownRenderer {
-	private _notebookURI: URI;
-	private _baseUrls: string[] = [];
-
-	constructor(
-		@IOpenerService private readonly _openerService: IOpenerService
-	) {
-	}
-
-	private getOptions(disposables: IDisposable[]): RenderOptions {
-		return {
-			actionHandler: {
-				callback: (content) => {
-					let uri: URI | undefined;
-					try {
-						uri = URI.parse(content);
-					} catch {
-						// ignore
-					}
-					if (uri && this._openerService) {
-						this._openerService.open(uri).catch(onUnexpectedError);
-					}
-				},
-				disposeables: disposables
-			}
-		};
-	}
-
-	render(markdown: IMarkdownString): IMarkdownRenderResult {
-		let disposables: IDisposable[] = [];
-		const element: HTMLElement = markdown ? this.renderMarkdown(markdown, this.getOptions(disposables)) : document.createElement('span');
-		return {
-			element,
-			dispose: () => dispose(disposables)
-		};
-	}
-
-	createElement(options: RenderOptions): HTMLElement {
-		const tagName = options.inline ? 'span' : 'div';
-		const element = document.createElement(tagName);
-		if (options.className) {
-			element.className = options.className;
-		}
-		return element;
-	}
-
-	parse(text: string): any {
-		let data = JSON.parse(text);
-		data = revive(data, 0);
-		return data;
-	}
-
-	/**
-	 * Create html nodes for the given content element.
-	 * Adapted from {@link htmlContentRenderer}
-	 */
-	renderMarkdown(markdown: IMarkdownString, options: RenderOptions = {}): HTMLElement {
-		const element = this.createElement(options);
-
-		// signal to code-block render that the
-		// element has been created
-		let signalInnerHTML: () => void;
-		const withInnerHTML = new Promise(c => signalInnerHTML = c);
-
-		let notebookFolder = path.dirname(this._notebookURI.fsPath) + '/';
-		if (!this._baseUrls.includes(notebookFolder)) {
-			this._baseUrls.push(notebookFolder);
-		}
-		const renderer = new marked.Renderer({ baseUrl: notebookFolder });
-		renderer.image = (href: string, title: string, text: string) => {
-			href = this.cleanUrl(!markdown.isTrusted, notebookFolder, href);
-			let dimensions: string[] = [];
-			if (href) {
-				const splitted = href.split('|').map(s => s.trim());
-				href = splitted[0];
-				const parameters = splitted[1];
-				if (parameters) {
-					const heightFromParams = /height=(\d+)/.exec(parameters);
-					const widthFromParams = /width=(\d+)/.exec(parameters);
-					const height = heightFromParams ? heightFromParams[1] : '';
-					const width = widthFromParams ? widthFromParams[1] : '';
-					const widthIsFinite = isFinite(parseInt(width));
-					const heightIsFinite = isFinite(parseInt(height));
-					if (widthIsFinite) {
-						dimensions.push(`width="${width}"`);
-					}
-					if (heightIsFinite) {
-						dimensions.push(`height="${height}"`);
-					}
-				}
-			}
-			let attributes: string[] = [];
-			if (href) {
-				attributes.push(`src="${href}"`);
-			}
-			if (text) {
-				attributes.push(`alt="${text}"`);
-			}
-			if (title) {
-				attributes.push(`title="${title}"`);
-			}
-			if (dimensions.length) {
-				attributes = attributes.concat(dimensions);
-			}
-			return '<img ' + attributes.join(' ') + '>';
-		};
-		renderer.link = (href, title, text): string => {
-			href = this.cleanUrl(!markdown.isTrusted, notebookFolder, href);
-			if (href === null) {
-				return text;
-			}
-			// Remove markdown escapes. Workaround for https://github.com/chjj/marked/issues/829
-			if (href === text) { // raw link case
-				text = removeMarkdownEscapes(text);
-			}
-			title = removeMarkdownEscapes(title);
-			href = removeMarkdownEscapes(href);
-			if (
-				!href
-				|| !markdown.isTrusted
-				|| href.match(/^data:|javascript:/i)
-				|| href.match(/^command:(\/\/\/)?_workbench\.downloadResource/i)
-			) {
-				// drop the link
-				return text;
-
-			} else {
-				// HTML Encode href
-				href = href.replace(/&/g, '&amp;')
-					.replace(/</g, '&lt;')
-					.replace(/>/g, '&gt;')
-					.replace(/"/g, '&quot;')
-					.replace(/'/g, '&#39;');
-				return `<a href=${href} data-href="${href}" title="${title || href}">${text}</a>`;
-			}
-		};
-		renderer.paragraph = (text): string => {
-			return `<p>${text}</p>`;
-		};
-
-		if (options.codeBlockRenderer) {
-			renderer.code = (code, lang) => {
-				const value = options.codeBlockRenderer!(lang, code);
-				// when code-block rendering is async we return sync
-				// but update the node with the real result later.
-				const id = defaultGenerator.nextId();
-
-				const promise = value.then(strValue => {
-					withInnerHTML.then(e => {
-						const span = element.querySelector(`div[data-code="${id}"]`);
-						if (span) {
-							span.innerHTML = strValue;
-						}
-					}).catch(err => {
-						// ignore
-					});
-				});
-
-				if (options.codeBlockRenderCallback) {
-					promise.then(options.codeBlockRenderCallback);
-				}
-
-				return `<div class="code" data-code="${id}">${escape(code)}</div>`;
-			};
-		}
-
-		if (options.actionHandler) {
-			options.actionHandler.disposeables.push(DOM.addStandardDisposableListener(element, 'click', event => {
-				let target: HTMLElement | null = event.target;
-				if (target.tagName !== 'A') {
-					target = target.parentElement;
-					if (!target || target.tagName !== 'A') {
-						return;
-					}
-				}
-				try {
-					const href = target.dataset['href'];
-					if (href) {
-						options.actionHandler!.callback(href, event);
-					}
-				} catch (err) {
-					onUnexpectedError(err);
-				} finally {
-					event.preventDefault();
-				}
-			}));
-		}
-
-		const markedOptions: marked.MarkedOptions = {
-			sanitize: !markdown.isTrusted,
-			renderer,
-			baseUrl: notebookFolder
-		};
-
-		element.innerHTML = marked.parse(markdown.value, markedOptions);
-		signalInnerHTML!();
-
-		return element;
-	}
-
-	// This following methods have been adapted from marked.js
-	// Copyright (c) 2011-2014, Christopher Jeffrey (https://github.com/chjj/)
-	cleanUrl(sanitize, base, href) {
-		if (sanitize) {
-			let prot;
-			try {
-				prot = decodeURIComponent(unescape(href))
-					.replace(/[^\w:]/g, '')
-					.toLowerCase();
-			} catch (e) {
-				return null;
-			}
-			if (prot.indexOf('javascript:') === 0 || prot.indexOf('vbscript:') === 0 || prot.indexOf('data:') === 0) {
-				return null;
-			}
-		}
-		try {
-			if (URI.parse(href)) {
-				return href;
-			}
-		} catch {
-			// ignore
-		}
-		let originIndependentUrl = /^$|^[a-z][a-z0-9+.-]*:|^[?#]/i;
-		if (base && !originIndependentUrl.test(href) && !fs.existsSync(href)) {
-			href = this.resolveUrl(base, href);
-		}
-		try {
-			href = encodeURI(href).replace(/%25/g, '%');
-		} catch (e) {
-			return null;
-		}
-		return href;
-	}
-
-	resolveUrl(base: string, href: string) {
-		if (!this._baseUrls[' ' + base]) {
-			// we can ignore everything in base after the last slash of its path component,
-			// but we might need to add _that_
-			// https://tools.ietf.org/html/rfc3986#section-3
-			if (/^[^:]+:\/*[^/]*$/.test(base)) {
-				this._baseUrls[' ' + base] = base + '/';
-			} else {
-				// Remove trailing 'c's. /c*$/ is vulnerable to REDOS.
-				this._baseUrls[' ' + base] = base.replace(/c*$/, '');
-			}
-		}
-		base = this._baseUrls[' ' + base];
-
-		if (href.slice(0, 2) === '//') {
-			return base.replace(/:[\s\S]*/, ':') + href;
-		} else if (href.charAt(0) === '/') {
-			return base.replace(/(:\/*[^/]*)[\s\S]*/, '$1') + href;
-		} else {
-			return base + href;
-		}
-	}
-
-	// end marked.js adaptation
-
-	setNotebookURI(val: URI) {
-		this._notebookURI = val;
-	}
-
-
 }
